pyproject.toml
[project]
name = "crogs-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "g4f>=0.6.0.1",
    "googletrans>=4.0.2",
    "pytelegrambotapi>=4.28.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "schedule>=1.2.2",
    "taskipy>=1.14.1",
]


[dependency-groups]
dev = [
    "watchfiles>=1.1.0",
    "pre-commit>=4.3.0",
    "ruff>=0.12.8",
]

[tool.taskipy.tasks] # uv run task <name>
prod = "python main.py --mode prod"
dev = "watchfiles 'python main.py --mode dev' --filter python  --ignore-paths __pycache__ --sigint-timeout 5"


[tool.ruff]
extend-include = ["*.ipynb"]
fix=true
unsafe-fixes=true
line-length=90

[tool.ruff.lint]
extend-select = ["I"]
ignore=["E402"]

[tool.pylint.messages_control]
disable = [
    "missing-function-docstring",
    "missing-module-docstring",
    "broad-exception-caught"
]
